requirements.txt
Babel==2.3.3
Django==1.5.12
Fabric==1.10.2
Fiona==1.5.1
GDAL==1.10.0
ModestMaps==1.4.6
Pillow==2.9.0
Shapely==1.2.16
South==1.0.2
Werkzeug==0.11.7
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
behave==1.2.5
billiard==3.3.0.20
boto==2.38.0
celery==3.1.18
certifi==2016.2.28
click==6.6
cligj==0.4.0
colorlog==2.6.0
coverage==3.6
crontab==0.20
cuisine==0.7.10
decorator==4.0.9
django-braces==1.8.1
django-csv-importer==0.1.3.5
django-data-tools==0.1
django-dotenv==1.4.1
django-draft==0.2.2
django-extensions==1.1.1
django-grappelli==2.4.12
django-guardian==1.2.5
django-jsonify==0.2.1
django-memoize==1.2.0
django-model-utils==2.0
django-nose==1.2
django-oauth-toolkit==0.9.0
django-picklefield==0.2.1
django-pytest==0.2.0
django-reversion==1.7
django-reversion-compare==0.3.5
django-sendfile==0.3.10
django-tastypie==0.9.14
ecdsa==0.13
enum34==1.1.3
fabtools==0.19.0
flake8==2.5.4
flower==0.8.3
funcsigs==1.0.0
futures==3.0.5
geojson==1.2.2
gunicorn==0.17.4
inflection==0.2.0
ipython==4.0.0
ipython-genutils==0.1.0
jsonfield==1.0.3
kombu==3.0.28
mccabe==0.4.0
meld3==1.0.2
mimeparse==0.1.3
mock==1.3.0
newrelic==2.66.0.49
nodeenv==0.6.5
nose==1.3.7
oauthlib==1.0.1
ordereddict==1.1
paramiko==1.16.0
parse==1.6.6
parse-type==0.3.4
pathlib2==2.1.0
pbr==1.9.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.2
pipdeptree==0.6.0
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
pycrypto==2.6.1
pyflakes==0.7.2
pytest==2.8.2
python-dateutil==2.5.2
python-memcached==1.48
pytz==2015.6
redis==2.10.3
requests==2.7.0
sarge==0.1.4
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sphinx-me==0.1.2
superlance==0.11
supervisor==3.2.3
testfixtures==4.7.0
tornado==4.3
traitlets==4.2.1
tzlocal==1.2
uuid==1.30
wsgiref==0.1.2

git+https://github.com/UrbanFootprint/TileStache.git@v1.50.3#egg=TileStache
git+https://github.com/UrbanFootprint/django-multi-gtfs.git@v0.3.3#egg=multigtfs
git+https://github.com/UrbanFootprint/django-paginator-rawqueryset.git@v1.0.0#egg=django-paginator-rawqueryset
git+https://github.com/UrbanFootprint/pyprof2calltree.git@v1.3.2#egg=pyprof2calltree
